requirements.txt
uvicorn
fastapi
python-dotenv
slowapi
pydantic
httpx
python-jose
google-genai